pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = 'kgcpy'
version = "1.1.5"
description = "Aids in identifying the Koeppen-Geiger (KG) climatic zone for a given latitude and longitude of any location."

readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    'pillow',
    'pandas',
]

license = {file = "LICENSE"}

classifiers = [
    'Development Status :: 4 - Beta',
    'License :: OSI Approved :: BSD License',
    'Operating System :: OS Independent',
    'Intended Audience :: Science/Research',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Topic :: Scientific/Engineering',
]

authors = [
  {name = "Xuanji Yu", email = "xuanjiy@bu.edu"},
  {name = "Chelsey Bryant", email = "clb117@case.edu"},
  {name = "Nicholas R. Wheeler", email = "nrw16@case.edu"},
  {name = "Franz Rubel", email = "franz.rubel@vetmeduni.ac.at"},
  {name = "Julian Ascencio Vasquez", email = "julian.ascencio@envision-digital.com"},
  {name = "Roger H. French", email = "rxf131@case.edu"},
]

[project.optional-dependencies]
test = ["pytest"]


[project.urls]
"Bug Tracker" = "https://github.com/cwru-sdle/kgcpy/issues"
"Source Code" = "https://github.com/cwru-sdle/kgcpy"
"Documentation" = "https://kgcpy.readthedocs.io"

[tool.setuptools.packages.find]
include = ["kgcpy*"]

[tool.setuptools_scm]
